password/requirements.txt
argon2-cffi
zxcvbn
fastapi
uvicorn
jinja2
python-multipart
itsdangerous
